requirements.txt
discord.py
Pillow
psutil
requests

# Testing
pytest
pytest-asyncio
pytest-mock
